+++ src/app/service/bounty/RefreshBounty.ts
@@ -2,22 +2,22 @@ import { GuildMember, Message, MessageEmbed } from 'discord.js';
 import { BountyCollection } from '../../types/bounty/BountyCollection';
 import { addPublishReactions } from './create/PublishBounty';
 import mongo, { Db } from 'mongodb';
-import dbInstance from '../../utils/dbUtils';
 import constants from '../constants/constants';
 import { addClaimReactions } from './ClaimBounty';
 import { addSubmitReactions } from './SubmitBounty';
 import { addCompletedReactions } from './CompleteBounty';
 import BountyUtils from '../../utils/BountyUtils';
 import Log from '../../utils/Log';
+import MongoDbUtils from '../../utils/MongoDbUtils';
 
 /**
  * This service will refresh the bounty in the Bounty board with the correct information
  * @param guildMember
  * @param bountyId
- * @param message 
+ * @param message
  */
 export default async (guildMember: GuildMember, bountyId: string, message: Message): Promise<any> => {
-	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_BOUNTY_BOARD);
+	const db: Db = await MongoDbUtils.connect(constants.DB_NAME_BOUNTY_BOARD);
 	const dbCollection = db.collection(constants.DB_COLLECTION_BOUNTIES);
 
 	const bountyCollection: BountyCollection = await dbCollection.findOne({