+++ src/app/types/poap/POAPTwitterParticipants.ts
@@ -1,0 +1,8 @@
+import { Collection } from 'mongodb';
+
+export interface POAPTwitterParticipants extends Collection {
+	twitterUserId: string,
+	twitterSpaceId: string,
+	startTime: string,
+	endTime: string,
+}